add cookstyle hook
Precommit-Verified: e76d070d4d6c118abee1c9c0783a78a3645e7d498b72785e6eb70a261867fb06
