+++ Gemfile
@@ -2,4 +2,4 @@
 
 source 'https://rubygems.org'
 
-gem 'rubocop', '0.57.2'
+gem 'rubocop', '0.77.0'
